[all] support for OT Posix nodes with RCP using realtime UART
This adds support for running a standard OT Posix node in a realtime simulation, where
the Posix process performs realtime SPINEL/HDLC UART communication with an RCP.
Support for RCPs is the basis to for upcoming features like including real OTBRs in a
simulation, either running on the local host or running in a Docker.

The dispatcher/simulation core is updated to allow handling of both 'realtime UART'
(stdout) based CLI and logging interactions and 'virtual-time UART' interactions by
nodes in the same simulation. The real-time simulation mode adapts the simulation
speed slightly, around speed=~1, to keep the time drift w.r.t. the system clock
near 0.

A method `waitForSimulation()` is introduced to avoid the need for custom simulation
event processing loops at different places in the code. Also, the queueing and
filtering of UART/log output of OT nodes is simplified to make the code better
readable and the data flow easier to understand and debug.

The CLI (over PTY/UART/...) echo behavior differs for Linux Posix nodes that are
controlled via stdin/stdout: there is no 'echo' behavior. Depending on OS, there could
be. To handle all possible combinations of platform/node-type, there is a function
'setupCli' added that autodetects whether echo is used for a node, or not, and whether
the CLI connection is functional.
To achieve this feature, some stale/unnecessary code related to the 'assurePrompt()'
function was removed, also making the node CLI code more maintainable.

The 'openthread' submodule is bumped to the latest version to get some required OT
new features in.

+++ pylibs/otns/cli/OTNS.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-# Copyright (c) 2020-2024, The OTNS Authors.
+# Copyright (c) 2020-2026, The OTNS Authors.
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -262,7 +262,7 @@ class OTNS(object):
         logging.basicConfig(level=level, format='%(asctime)s - %(levelname)s - %(message)s')
 
     @property
-    def time(self) -> int:
+    def time(self) -> float:
         """
         Get simulation time.
